requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
urllib3>=2.0.7
certifi
playwright>=1.40.0
lxml>=4.9.0
